feat: update vapiCallHistory model and add unique constraint to callId
feat: modify dietRouter to handle user diet plan retrieval

feat: create eventRouter for handling server events

feat: update exerciseRouter to remove unused endpoint

feat: implement server event service for client management

feat: enhance webhook services to process vapi call history and integrate diet/workout plan generation

feat: create utility functions for generating diet and workout plans

style: add new styles for day program and selector components

feat: implement DayProgram component to display workout and diet plans

fix: adjust ProtectedRoute component for better responsiveness

fix: disable StrictMode in main.tsx for compatibility

feat: enhance Dashboard to fetch and display user workout and diet plans

feat: refactor GenerateProgram to handle voice call interactions and improve UI

fix: update Home page to hide navigation links

fix: adjust Profile page for better layout

feat: create utility functions for fetching user plans

+++ backend/src/services/webhookServices.js
@@ -1,9 +1,9 @@
-import { Webhook } from "svix";
-import { ENV } from "../lib/env.js";
 import { processedEvent } from "../models/webhookEventModel.js";
 import { userModel } from "../models/userModel.js";
-import {AppError} from '../utils/AppErrorHandler.js'
-
+import { vapiCallHistoryModel } from "../models/vapiCallHistoryModel.js";
+import { AppError, generateDietAndWorkoutPlan } from "../utils/index.js";
+import { handleAddUserNewExPlan } from "../controllers/exerciseController.js";
+import { handleAddUserNewDietPlan } from "../controllers/dietController.js";
 
 const isalreadyProcessed = async ({ eventId, eventType, eventSource }) => {
   try {
@@ -19,32 +19,72 @@ const isalreadyProcessed = async ({ eventId, eventType, eventSource }) => {
   }
 };
 
-export const processClerkWebhook = async (eventId, event) => {
+export const processWebhook = async (eventId, event, eventSource) => {
   try {
     const shouldProcess = await isalreadyProcessed({
       eventId,
       eventType: event.type,
-      eventSource: "clerk",
+      eventSource,
     });
     if (shouldProcess) {
-      switch (event.type) {
-        case "user.created":
-          await userModel.create({
-            clerkUserId: event.data.id,
-          });
-          break;
-        case "user.deleted":
-          await userModel.findOneAndDelete({ clerkUserId: event.data.id });
+      if (eventSource === "clerk") {
+        switch (event.type) {
+          case "user.created":
+            await userModel.create({
+              clerkUserId: event.data.id,
+            });
+            break;
+          case "user.deleted":
+            await userModel.findOneAndDelete({ clerkUserId: event.data.id });
 
-          break;
+            break;
 
-        default:
-          console.log(`Unhandled event type: ${event.type}`);
+          default:
+            console.log(`Unhandled event type: ${event.type}`);
+        }
+        return "clerk webhook processed";
+      } else {
+        const structuredData = event.analysis.structuredData;
+        await vapiCallHistoryModel.create({
+          clerkUserId:
+            event.call?.assistantOverrides.variableValues.clerkUserId,
+          callId: event.call.id,
+          intake: {
+            ...structuredData,
+          },
+          rawStructuredData:event
+        });
+        if (
+          !Object.values(structuredData).some(
+            (i) => i === "" || i === null || i === undefined,
+          )
+        ) {
+          const result = await generateDietAndWorkoutPlan(structuredData);
+          await Promise.all([
+            handleAddUserNewExPlan({
+              clerkUserId:
+                event.call?.assistantOverrides.variableValues.clerkUserId,
+              vapiCallId: event.call?.id,
+              status: "active",
+              daysPerWeek: structuredData.gymDaysPerWeek,
+              ...result.exercisePlan,
+            }),
+            handleAddUserNewDietPlan({
+              clerkUserId:
+                event.call?.assistantOverrides.variableValues.clerkUserId,
+              vapiCallId: event.call?.id,
+              status: "active",
+              ...result["dietPlan"],
+            }),
+          ]);
+          sendUserPlan(event.call?.assistantOverrides.variableValues.clerkUserId)
+        }
+
+        return "vapi webhook processed";
       }
-
     }
   } catch (err) {
-    console.error("Error processing clerk webhook:", err.message);
+    console.error(`Error processing ${eventSource} webhook:`, err.message);
     // Return 500 so Clerk retries later
     throw new AppError("Processing failed", 500);
   }
